Refactoring ci-dashboard-backend code
- Add a new folder `src` which contains all platform specific functions
- rename `server.js` file to `main.js` file
- Modify launch.sh for staring `main.js` file
