+++ launch.sh
@@ -2,6 +2,6 @@
 
 echo "Generating keys for server"
 
-token=$KEY node server.js
+token=$KEY node main.js
 
 echo "Starting server"
